FastAPIBackend/requirements.txt
jinja2
fastapi
uvicorn
pirebase
gunicorn
pymongo
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
ipython
ipykernel
numpy
pandas 
matplotlib
python-dotenv
httpx